+++ game/king.py
@@ -1,6 +1,6 @@
 import sys
 import os
-sys.path.append(os.path.realpath(os.path.join(os.path.dirname(__file__), '..')))
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
 
 from game.piece import Piece
 
@@ -22,7 +22,7 @@ class King(Piece):
         col_diff = abs(end_col - start_col) # Calcula la diferencia entre la columna final y la columna inicial
 
         # El rey puede moverse una casilla en cualquier dirección
-        if row_diff <= 1 and col_diff <= 1:  # Verifica si la diferencia que se movió es menor o igual a 1 en fila y columna
+        if max(row_diff, col_diff) == 1:  # Verifica si la diferencia máxima en fila o columna es 1
             destination_piece = board[end_row][end_col]
             if destination_piece is None or destination_piece.color != self.color:  # Verificar si la casilla de destino está vacía o contiene una pieza del oponente
                 return True