+++ game/knight.py
@@ -22,8 +22,11 @@ class Knight(Piece):
         col_diff = abs(end_col - start_col) # Calcula la diferencia entre la columna final y la columna inicial
 
         # El caballo se mueve en forma de "L": dos casillas en una dirección y una en la otra
-        if (row_diff, col_diff) in [(2, 1), (1, 2)]:  # Verifica si el movimiento es en forma de "L"
+        is_knight_move = (row_diff == 2 and col_diff == 1) or (row_diff == 1 and col_diff == 2)
+
+        if is_knight_move:
             destination_piece = board[end_row][end_col]
-            if destination_piece is None or destination_piece.color != self.color:  # Verificar si la casilla de destino está vacía o contiene una pieza del oponente
+            is_empty_or_enemy = destination_piece is None or destination_piece.color != self.color
+            if is_empty_or_enemy:
                 return True
         return False